fix: Make Prisma schema test environment-aware
- Update basic test to accept both SQLite (dev) and PostgreSQL (prod) providers
- Fix pre-commit test failure that was blocking commits
- Allow both database configurations as valid
- All 11 basic tests now pass
