+++ prisma/schema.prisma
@@ -6,7 +6,7 @@ generator client {
 }
 
 datasource db {
-  provider = "postgresql"
+  provider = "sqlite"
   url      = env("DATABASE_URL")
 }
 
@@ -118,7 +118,7 @@ model Integration {
   expiresAt    DateTime?
   isActive     Boolean @default(true)
   lastSyncAt   DateTime?
-  metadata     Json @default("{}")
+  metadata     String @default("{}")
   
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
